refactor: convert to Long Type

+++ src/main/java/com/d108/project/domain/businessArea/repository/BusinessAreaRepository.java
@@ -3,5 +3,5 @@ package com.d108.project.domain.businessArea.repository;
 import com.d108.project.domain.businessArea.entity.BusinessArea;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BusinessAreaRepository extends JpaRepository<BusinessArea, Integer> {
+public interface BusinessAreaRepository extends JpaRepository<BusinessArea, Long> {
 }
