+++ src/main/java/com/d108/project/domain/favorite/favoriteFranchise/service/FavoriteFranchiseService.java
@@ -6,7 +6,7 @@ import com.d108.project.domain.member.entity.Member;
 import java.util.List;
 
 public interface FavoriteFranchiseService {
-    FavoriteFranchise createFavoriteFranchise(Member member, Integer franchiseId);
-    void deleteFavoriteFranchise(Integer id);
+    FavoriteFranchise createFavoriteFranchise(Member member, Long franchiseId);
+    void deleteFavoriteFranchise(Long id);
     List<FavoriteFranchise> getFavoriteFranchisesByMember(Member member);
 }
