+++ src/main/java/com/d108/project/domain/forum/saleStorePost/repository/SaleStorePostRepository.java
@@ -3,5 +3,5 @@ package com.d108.project.domain.forum.saleStorePost.repository;
 import com.d108.project.domain.forum.saleStorePost.entity.SaleStorePost;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SaleStorePostRepository extends JpaRepository<SaleStorePost, Integer> {
+public interface SaleStorePostRepository extends JpaRepository<SaleStorePost, Long> {
 }
