+++ src/main/java/com/d108/project/domain/logincredential/repository/LoginCredentialRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface LoginCredentialRepository extends JpaRepository<LoginCredential, Integer> {
+public interface LoginCredentialRepository extends JpaRepository<LoginCredential, Long> {
     Optional<LoginCredential> findByUsername(String username);
 }
